% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_from_data.R
\name{genQueryFilters}
\alias{genQueryFilters}
\title{Generate filters definition}
\usage{
genQueryFilters(
  data,
  settings = list(),
  .queryBuilderConfig = queryBuilder::queryBuilderConfig
)
}
\arguments{
\item{data}{Dataset from which filters should extracted.}

\item{settings}{Named list. Column-specific filter configuration.
For each variable the provided settings will overwrite the default ones.}

\item{.queryBuilderConfig}{R6 object of class 'queryBuilderConfig' storing queryOperators.
See \link[queryBuilder]{query-operator}.}
}
\value{
Nested list object storing generated filters configuration.
}
\description{
Generate filters definition
}
\examples{

genQueryFilters(
  iris,
  list(
    Species = list(operators = c("equal", "not_equal"))
  )
)

}
