% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_input.R
\name{js}
\alias{js}
\title{Store JS definition as character string}
\usage{
js(x)
}
\arguments{
\item{x}{Character string containing valid JS object e.g. function}
}
\value{
An object of class 'json' storing the provided character string.
}
\description{
Store JS definition as character string
}
