% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{r_to_js_opt_type}
\alias{r_to_js_opt_type}
\title{Convert R class to a valid operator JS type}
\usage{
r_to_js_opt_type(apply_to)
}
\arguments{
\item{apply_to}{Character value - R class to be converted.}
}
\description{
Convert R class to a valid operator JS type
}
